usb-dwc/stm32: relocate common otg_fs, otg_hs implementations as both STM32 and EFM32 chips use the same dwc USB IP cores
