+++ lib/usb/usb_f107.c
@@ -20,9 +20,9 @@
 #include <string.h>
 #include <libopencm3/cm3/common.h>
 #include <libopencm3/stm32/tools.h>
-#include <libopencm3/stm32/otg_fs.h>
 #include <libopencm3/stm32/rcc.h>
 #include <libopencm3/usb/usbd.h>
+#include <libopencm3/usb/dwc/otg_fs.h>
 #include "usb_private.h"
 #include "usb_dwc_common.h"
